refactor(dsl-protobuf): tighten builder and support boundaries
