+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/field/OneofFieldBuilder.kt
@@ -2,7 +2,7 @@ package me.liam.microsmith.dsl.schemas.protobuf.field
 
 import me.liam.microsmith.dsl.schemas.protobuf.OneofFieldScope
 
-class OneofFieldBuilder(var index: Int? = null) : OneofFieldScope {
+internal class OneofFieldBuilder(var index: Int? = null) : OneofFieldScope {
     override fun index(index: Int) {
         this.index = index
     }
