+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/field/ReferenceFieldBuilder.kt
@@ -2,7 +2,7 @@ package me.liam.microsmith.dsl.schemas.protobuf.field
 
 import me.liam.microsmith.dsl.schemas.protobuf.ReferenceFieldScope
 
-class ReferenceFieldBuilder(var index: Int? = null, var cardinality: Cardinality = Cardinality.REQUIRED) :
+internal class ReferenceFieldBuilder(var index: Int? = null, var cardinality: Cardinality = Cardinality.REQUIRED) :
     ReferenceFieldScope {
     override fun index(index: Int) {
         this.index = index
