+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/field/ScalarFieldBuilder.kt
@@ -2,7 +2,7 @@ package me.liam.microsmith.dsl.schemas.protobuf.field
 
 import me.liam.microsmith.dsl.schemas.protobuf.ScalarFieldScope
 
-class ScalarFieldBuilder(var index: Int? = null, var cardinality: Cardinality = Cardinality.REQUIRED) :
+internal class ScalarFieldBuilder(var index: Int? = null, var cardinality: Cardinality = Cardinality.REQUIRED) :
     ScalarFieldScope {
     override fun optional() {
         require(cardinality == Cardinality.REQUIRED) {
