+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/reserved/ReservedDeclarations.kt
@@ -1,0 +1,10 @@
+package me.liam.microsmith.dsl.schemas.protobuf.reserved
+
+import me.liam.microsmith.dsl.schemas.protobuf.support.IndexAllocator
+import me.liam.microsmith.dsl.schemas.protobuf.support.NameRegistry
+
+internal fun buildReservedDeclarations(allocator: IndexAllocator, nameRegistry: NameRegistry): List<Reserved> =
+    buildList {
+        allocator.reserved().sortedBy(IntRange::first).mapTo(this, Reserved::fromRange)
+        nameRegistry.reserved().sorted().mapTo(this, ::ReservedName)
+    }
