+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/support/IndexAllocator.kt
@@ -3,7 +3,7 @@ package me.liam.microsmith.dsl.schemas.protobuf.support
 import me.liam.microsmith.dsl.schemas.protobuf.extensions.merge
 import me.liam.microsmith.dsl.schemas.protobuf.reserved.Max
 
-class IndexAllocator(private val min: Int, private val protoReserved: IntRange? = null) {
+internal class IndexAllocator(private val min: Int, private val protoReserved: IntRange? = null) {
     private val reserved = mutableSetOf<IntRange>()
 
     fun reserved() = reserved.toSet()
