+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/types/EnumValueBuilder.kt
@@ -2,7 +2,7 @@ package me.liam.microsmith.dsl.schemas.protobuf.types
 
 import me.liam.microsmith.dsl.schemas.protobuf.EnumValueScope
 
-class EnumValueBuilder(var index: Int? = null) : EnumValueScope {
+internal class EnumValueBuilder(var index: Int? = null) : EnumValueScope {
     override fun index(index: Int) {
         this.index = index
     }
